backend/requirements.txt
# PRISMA Backend Dependencies
# Python 3.10+

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# HTTP Client & API Integration
# ============================================================================
requests==2.31.0
httpx==0.25.1  # Alternative async HTTP client

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.1.3
numpy==1.26.2

# ============================================================================
# Web Scraping (for web_enrichment module)
# ============================================================================
beautifulsoup4==4.12.2
lxml==4.9.3

# ============================================================================
# File Parsing (CSV/Excel upload)
# ============================================================================
openpyxl==3.1.2  # Excel file support
python-multipart==0.0.6  # File upload support

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# LLM Integration (for future llm/engine.py)
# ============================================================================
# Note: Ollama doesn't need a Python package, we use HTTP requests
# But if we want a wrapper:
# ollama-python==0.1.0  # Optional convenience wrapper

# ============================================================================
# Date/Time Utilities
# ============================================================================
python-dateutil==2.8.2

# ============================================================================
# Testing (Optional but recommended)
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1  # For testing FastAPI endpoints

# ============================================================================
# Development Tools (Optional)
# ============================================================================
black==23.11.0  # Code formatter
flake8==6.1.0   # Linter
mypy==1.7.1     # Type checker

# ============================================================================
# Database (Optional - for production)
# ============================================================================
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # PostgreSQL
# alembic==1.12.1  # Database migrations

